Fix macOS alias resolution and type conversions
- switch alias resolution to NSURL APIs with autorelease pool
- match IOHIDSystem service by class name string to avoid header dependency
- cast window layer to int32 for platform info struct

+++ window/list_windows_darwin.go
@@ -216,7 +216,7 @@ func listWindows(options WindowOptions) ([]WindowInfo, error) {
 			platform: &DarwinPlatformInfo{
 				WindowID:  uint32(item.windowID),
 				OwnerName: ownerName,
-				Layer:     item.layer,
+				Layer:     int32(item.layer),
 				Alpha:     float64(item.alpha),
 				Onscreen:  onScreen,
 			},
